Routes folder created, all routes in seperate folders
