+++ images/api/src/routes/users.js
@@ -1,0 +1,19 @@
+const express = require('express');
+const knex = require('knex');
+const knexfile = require('../knexfile');
+const {
+    v4: uuidv4
+} = require('uuid');
+const router = express.Router();
+const PORT = 3000;
+
+const db = knex(knexfile.development);
+
+const bodyParser = require('body-parser');
+router.use(bodyParser.json());
+router.use(express.json());
+router.use(bodyParser.json());
+
+
+
+module.exports = router;
